refactoring render functions were moved to renderes.js
